Double and remove plookup_composer.rs
